Rename config.ts to types.ts to avoid confusion
