+++ packages/core/src/generate/config/index.ts
@@ -1,2 +1,2 @@
-export * from "./config";
+export * from "./types";
 export * from "./parseAppConfig";
