+++ packages/core/src/generate/config/types.ts
